test: improve and fix flaky
