+++ tests/utils/actions/misc.ts
@@ -9,7 +9,6 @@ import type { RichText } from '../../../packages/playground/examples/virgo/test-
 import type { BaseBlockModel } from '../../../packages/store/src/index.js';
 import {
   pressEnter,
-  pressEscape,
   pressSpace,
   pressTab,
   SHORT_KEY,
@@ -307,22 +306,6 @@ export async function initEmptyDatabaseWithParagraphState(
     return { pageId, frameId, databaseId };
   }, pageId);
   return ids;
-}
-
-export async function initDatabaseColumn(page: Page, title = '') {
-  const header = page.locator('.affine-database-column-header');
-  const box = await getBoundingBox(header);
-  await page.mouse.move(box.x + box.width / 2, box.y + box.height / 2);
-
-  const columnAddBtn = page.locator('.header-add-column-button');
-  await columnAddBtn.click();
-
-  if (title) {
-    await type(page, title);
-    await pressEnter(page);
-  } else {
-    await pressEscape(page);
-  }
 }
 
 export async function initDatabaseRow(page: Page) {
